chatting & pay & product & login MERGE DRAFT
