+++ package.json
@@ -7,32 +7,34 @@
     "clean": "rm -rf build public/bundle.js",
     "build": "babel server --out-dir build && webpack",
     "start": "SOCKET_URL=localhost:3000 BABEL_DISABLE_CACHE=1 NODE_ENV=development node ./build/main.js",
-    "start:dev": "NODE_ENV=production node ./build/main.js",
+    "start:prod": "NODE_ENV=production node ./build/main.js",
     "test": "webpack-dev-server --config webpack.dev.config.js"
   },
   "author": "velopert <public.velopert@gmail.com> (https://velopert.com)",
   "license": "ISC",
   "dependencies": {
     "express": "^4.13.4",
+    "file-loader": "^0.10.0",
+    "iamport": "^0.3.4",
+    "importer": "^0.7.2",
+    "jquery": "^3.1.1",
+    "material-ui": "^0.17.0",
     "mongoose": "^4.4",
+    "passport": "^0.3.2",
+    "passport-facebook": "^2.1.1",
+    "passport-local": "^1.0.0",
     "q": "^1.4.1",
     "react": "^15.4.2",
     "react-dom": "^15.4.2",
-    "react-tap-event-plugin": "^2.0.1",
-    "react-facebook-login": "^3.4.3",
-    "iamport": "^0.3.4",
-    "importer": "^0.7.2",
-    "file-loader": "^0.10.0",
-    "jquery": "^3.1.1",
-    "material-ui": "^0.17.0",
     "react-dropzone": "^3.10.0",
+    "react-facebook-login": "^3.4.3",
     "react-redux": "^5.0.2",
-    "redux": "^3.6.0",
     "react-router": "^3.0.2",
+    "react-tap-event-plugin": "^2.0.1",
+    "redux": "^3.6.0",
     "socket.io": "^1.7.2",
     "socket.io-client": "^1.7.3",
     "socket.io-react": "^1.1.2"
-
   },
   "devDependencies": {
     "babel-cli": "^6.9.0",
@@ -45,6 +47,5 @@
     "react-hot-loader": "^1.3.0",
     "webpack": "^1.13.0",
     "webpack-dev-server": "^1.14.1"
-    
   }
 }
