+++ src/components/App.js
@@ -1,30 +1,6 @@
 import React from 'react';
-
-import Home from './Home';
 import { Router, Route, browserHistory } from 'react-router';
-import Facebook from './Facebook';
-import Welcome from './Welcome';
 
-import BuyerLogin from './BuyerLogin';
-import SellerLogin from './SellerLogin';
-import Join from './Join';
-
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
-import Addproducts from './Addproducts';
-import ProductLists from './ProductLists';
-import Billbox from './billbox';
-import KaKao from './kakao';
-
-import ChatBox from './chatbox/ChatBox.js';
-
-// react router
-import { Link } from 'react-router'
-
-//socket
-import { SocketProvider } from 'socket.io-react';
-import io from 'socket.io-client';
-const socket = io.connect(process.env.SOCKET_URL);
-			socket.on('connectMsg', msg => console.log(msg));
 
 class App extends React.Component {
 
@@ -35,34 +11,17 @@ class App extends React.Component {
     render() {
 
         return (
-        	
-          <div>
-            <Addproducts></Addproducts>
-            <ProductLists></ProductLists>
-            <Billbox></Billbox>
-
-
-						<ul><li><Link to="intro">Intro</Link></li><li><Link to="aaaaa">NotFound</Link></li></ul>
-            	{this.props.children}
-            <h1>This is Bj Shop Mall</h1>
-						<SocketProvider socket={socket}>
-							<ChatBox user='chan'/>
-						</SocketProvider>	 
-          </div>
-
-        	<Router history={browserHistory}>
-        		<Route path="/" component={Home}></Route>
-        		<Route path="/auth/facebook" component={Facebook}></Route>
-        		<Route path="/welcome" component={Welcome}></Route>
-        		<Route path="/sellerLogin" component={SellerLogin}></Route>
-        		<Route path="/buyerLogin" component={BuyerLogin}></Route>
-        		<Route path="/join" component={Join}></Route>
-        	</Router>
-
+        	<div>
+        		{this.props.children}
+        	</div>
         )
     }
 }
 
 export default App;
+
+
+// 나머지 path를 제외한 모든 경로는 * path로 이동
+// Route path= "*" component = {}
